+++ services/vision/mlvision/ml_model.go
@@ -3,7 +3,10 @@
 package mlvision
 
 import (
+	"bufio"
 	"context"
+	"os"
+	"strings"
 
 	"github.com/edaniels/golog"
 	"github.com/pkg/errors"
@@ -19,6 +22,13 @@ import (
 )
 
 var model = resource.NewDefaultModel("mlmodel")
+
+const (
+	// UInt8 is one of the possible input/output types for tensors.
+	UInt8 = "uint8"
+	// Float32 is one of the possible input/output types for tensors.
+	Float32 = "float32"
+)
 
 func init() {
 	registry.RegisterService(vision.Subtype, model, registry.Service{
@@ -82,4 +92,88 @@ func registerMLModelVisionService(
 	}
 	// Don't return a close function, because you don't want to close the underlying ML service
 	return vision.NewService(name, r, nil, classifierFunc, detectorFunc, segmenter3DFunc)
+}
+
+// Unpack output based on expected type and force it into a []float64.
+func unpack(inMap map[string]interface{}, name string, md mlmodel.MLMetadata) []float64 {
+	var out []float64
+	me := inMap[name]
+	switch getTensorTypeFromName(name, md) {
+	case UInt8:
+		temp := me.([]uint8)
+		for _, t := range temp {
+			out = append(out, float64(t))
+		}
+	case Float32:
+		temp := me.([]float32)
+		for _, p := range temp {
+			out = append(out, float64(p))
+		}
+	default:
+		return nil
+	}
+	return out
+}
+
+func getTensorTypeFromName(name string, md mlmodel.MLMetadata) string {
+	for _, o := range md.Outputs {
+		if strings.Contains(strings.ToLower(o.Name), strings.ToLower(name)) {
+			return o.DataType
+		}
+	}
+	return ""
+}
+
+// getLabelsFromMetadata returns a slice of strings--the intended labels
+func getLabelsFromMetadata(md mlmodel.MLMetadata) []string {
+	for _, o := range md.Outputs {
+		if strings.Contains(o.Name, "category") || strings.Contains(o.Name, "probability") {
+			if labelPath, ok := o.Extra["labels"]; ok {
+				labels := []string{}
+				f, err := os.Open(labelPath.(string))
+				if err != nil {
+					return nil
+				}
+				defer func() {
+					if err := f.Close(); err != nil {
+						panic(err)
+					}
+				}()
+				scanner := bufio.NewScanner(f)
+				for scanner.Scan() {
+					labels = append(labels, scanner.Text())
+				}
+				return labels
+			}
+		}
+	}
+	return nil
+}
+
+// getBoxOrderFromMetadata returns a slice of ints--the bounding box
+// display order, where 0=xmin, 1=ymin, 2=xmax, 3=ymax
+func getBoxOrderFromMetadata(md mlmodel.MLMetadata) ([]int, error) {
+	for _, o := range md.Outputs {
+		if strings.Contains(o.Name, "location") {
+			out := make([]int, 0, 4)
+			if order, ok := o.Extra["boxOrder"].([]uint32); ok {
+				for _, o := range order {
+					out = append(out, int(o))
+				}
+				return out, nil
+			}
+		}
+	}
+	return nil, errors.New("could not grab bbox order")
+}
+
+// getIndex returns the index of an int in an array of ints
+// Will return -1 if it's not there.
+func getIndex(s []int, num int) int {
+	for i, v := range s {
+		if v == num {
+			return i
+		}
+	}
+	return -1
 }
